Version 1.21 (2012-04-25)

+++ examples/example-GeodesicLine.cpp
@@ -1,5 +1,5 @@
 // Example of using the GeographicLib::GeodesicLine class
-// $Id: 8105576ed7fb1946223d89d7394ada3ef06e7a93 $
+// $Id: df5c5ffe6534d46e343544430de8c13fc8119bfb $
 
 #include <iostream>
 #include <exception>
@@ -26,7 +26,7 @@ int main() {
     // Alternatively
     // const GeographicLib::GeodesicLine line = geod.Line(lat1, lon1, azi1);
     double ds = 500e3;          // Nominal distance between points = 500 km
-    int num = ceil(s12 / ds);   // The number of intervals
+    int num = int(ceil(s12 / ds)); // The number of intervals
     cout << fixed << setprecision(3);
     {
       // Use intervals of equal length
